+++ client/src/components/UserInfo/UserInfoTap.tsx
@@ -2,23 +2,26 @@ import React, { useState, useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import styled from 'styled-components';
 import FilterTap from '../common/FilterTap';
-import { userInfoTab, otherInfoTab } from '../../data/filterTapList';
+import { userInfoTab, otherInfoTab } from '../../data/filterTabList';
 import { StyledMainContent } from '../../pages/GameChannel';
 import SelectTag from '../common/SelectTag';
 import postOptionTags, { optionMapping } from '../../data/postOptionTags';
 
-const UserInfoTap = ({ isSameUser, setIsSelectTab, isSelectTab, setIsSelectTag, setIsMappingTag  }
-  : { 
-      isSameUser: boolean,
-      setIsSelectTab: React.Dispatch<React.SetStateAction<string>>,
-      isSelectTab: string,
-      setIsSelectTag: React.Dispatch<React.SetStateAction<string>>,
-      setIsMappingTag: React.Dispatch<React.SetStateAction<string>>,
-  }
-  ) => {
-
+const UserInfoTap = ({
+  isSameUser,
+  setIsSelectTab,
+  isSelectTab,
+  setIsSelectTag,
+  setIsMappingTag
+}: {
+  isSameUser: boolean;
+  setIsSelectTab: React.Dispatch<React.SetStateAction<string>>;
+  isSelectTab: string;
+  setIsSelectTag: React.Dispatch<React.SetStateAction<string>>;
+  setIsMappingTag: React.Dispatch<React.SetStateAction<string>>;
+}) => {
   const { memberId } = useParams();
-  const [ showSelectTag, setShowSelectTag ] = useState<boolean>(false);
+  const [showSelectTag, setShowSelectTag] = useState<boolean>(false);
 
   useEffect(() => {
     if (isSameUser) {
@@ -35,8 +38,8 @@ const UserInfoTap = ({ isSameUser, setIsSelectTab, isSelectTab, setIsSelectTag, 
   useEffect(() => {
     setShowSelectTag(
       isSelectTab === '북마크 글' ||
-      isSelectTab === '내가 쓴 글' ||
-      isSelectTab === '작성글'
+        isSelectTab === '내가 쓴 글' ||
+        isSelectTab === '작성글'
     );
   }, [isSelectTab]);
 
@@ -51,16 +54,13 @@ const UserInfoTap = ({ isSameUser, setIsSelectTab, isSelectTab, setIsSelectTag, 
 
   return (
     <StyledWrapper>
-      { showSelectTag && 
+      {showSelectTag && (
         <StyledSubContent>
-          <SelectTag 
-            options={postOptionTags}
-            onChange={handleChange}
-          />
+          <SelectTag options={postOptionTags} onChange={handleChange} />
         </StyledSubContent>
-      }
-      <FilterTap 
-        filterList={isSameUser ? userInfoTab : otherInfoTab} 
+      )}
+      <FilterTap
+        filterList={isSameUser ? userInfoTab : otherInfoTab}
         onClickFilter={handleClick}
       />
     </StyledWrapper>
